Make ssphys buildable on 64-bit Linux.
The 'long' type is 32-bit in MSVC++, and 64-bit on Linux,
so nobody should use it without good reason.

+++ ssphys/utils/tinyxml.h
@@ -355,7 +355,7 @@ protected:
 			return tolower( v );
 		}
 	}
-	static void ConvertUTF32ToUTF8( unsigned long input, char* output, int* length );
+	static void ConvertUTF32ToUTF8( unsigned input, char* output, int* length );
 
 private:
 	TiXmlBase( const TiXmlBase& );				// not implemented.
