+++ ssphys/utils/tinyxmlparser.cpp
@@ -84,11 +84,11 @@ const int TiXmlBase::utf8ByteTable[256] =
 };
 
 
-void TiXmlBase::ConvertUTF32ToUTF8( unsigned long input, char* output, int* length )
+void TiXmlBase::ConvertUTF32ToUTF8( unsigned input, char* output, int* length )
 {
-	const unsigned long BYTE_MASK = 0xBF;
-	const unsigned long BYTE_MARK = 0x80;
-	const unsigned long FIRST_BYTE_MARK[7] = { 0x00, 0x00, 0xC0, 0xE0, 0xF0, 0xF8, 0xFC };
+	const unsigned BYTE_MASK = 0xBF;
+	const unsigned BYTE_MARK = 0x80;
+	const unsigned FIRST_BYTE_MARK[7] = { 0x00, 0x00, 0xC0, 0xE0, 0xF0, 0xF8, 0xFC };
 
 	if (input < 0x80) 
 		*length = 1;
@@ -433,7 +433,7 @@ const char* TiXmlBase::GetEntity( const char* p, char* value, int* length, TiXml
 
 	if ( *(p+1) && *(p+1) == '#' && *(p+2) )
 	{
-		unsigned long ucs = 0;
+		unsigned ucs = 0;
 		ptrdiff_t delta = 0;
 		unsigned mult = 1;
